backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncmy==0.2.8
aiosqlite==0.19.0
PyMySQL==1.1.0
pydantic==2.5.0
python-multipart==0.0.6
openai==1.3.8
azure-storage-blob==12.19.0
python-dotenv==1.0.0
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
Pillow==10.0.1
pytz==2024.1
aiohttp==3.9.1